feat: create ability for users to follow creators they are buying (defaults to true
