+++ src/app/trade-creator-page/trade-creator-complete/trade-creator-complete.component.html
@@ -1,6 +1,9 @@
 <!-- Override .container's max-width property with max-width: inherit-->
 <div class="container fs-15px" style="max-width: inherit">
-  <trade-creator-table [creatorCoinTrade]="creatorCoinTrade"></trade-creator-table>
+  <trade-creator-table
+    [creatorCoinTrade]="creatorCoinTrade"
+    [userFollowingCreator]="userFollowingCreator"
+  ></trade-creator-table>
 
   <!-- Buttons -->
   <div class="row no-gutters pt-3">
