+++ src/app/trade-creator-page/trade-creator-complete/trade-creator-complete.component.ts
@@ -11,6 +11,7 @@ import { CreatorCoinTrade } from "../../../lib/trade-creator-page/creator-coin-t
 })
 export class TradeCreatorCompleteComponent implements OnInit {
   @Input() creatorCoinTrade: CreatorCoinTrade;
+  @Input() userFollowingCreator: boolean;
   @Output() tradeAgainButtonClicked = new EventEmitter();
 
   router: Router;
@@ -32,6 +33,8 @@ export class TradeCreatorCompleteComponent implements OnInit {
   }
 
   ngOnInit() {
+    console.log('Here is the user following creator')
+    console.log(this.userFollowingCreator);
     window.scroll(0, 0);
   }
 }
