+++ src/app/trade-creator-page/trade-creator-form/trade-creator-form.component.ts
@@ -1,4 +1,4 @@
-import { Component, EventEmitter, Input, OnDestroy, OnInit, Output } from "@angular/core";
+import {ChangeDetectorRef, Component, EventEmitter, Input, OnDestroy, OnInit, Output} from "@angular/core";
 import { Location } from "@angular/common";
 import { ActivatedRoute, Router } from "@angular/router";
 import { GlobalVarsService } from "../../global-vars.service";
@@ -31,6 +31,7 @@ export class TradeCreatorFormComponent implements OnInit, OnDestroy {
   MIN_BITCLOUT_NANOS_TO_LEAVE_WHEN_BUYING_CREATOR_COINS = 100_000;
 
   @Input() creatorCoinTrade: CreatorCoinTrade;
+  @Input() userFollowingCreator: boolean;
   @Output() previewClicked = new EventEmitter();
 
   router: Router;
