+++ src/app/trade-creator-page/trade-creator/trade-creator.component.html
@@ -63,6 +63,7 @@
     <div *ngIf="screenToShow == this.TRADE_CREATOR_FORM_SCREEN">
       <trade-creator-form
         [creatorCoinTrade]="creatorCoinTrade"
+        [userFollowingCreator]="userFollowingCreator"
         (previewClicked)="_onPreviewClicked()"
       ></trade-creator-form>
     </div>
@@ -70,6 +71,7 @@
     <div *ngIf="screenToShow == this.TRADE_CREATOR_PREVIEW_SCREEN">
       <trade-creator-preview
         [creatorCoinTrade]="creatorCoinTrade"
+        [userFollowingCreator]="userFollowingCreator"
         (slippageError)="_onSlippageError()"
         (backButtonClicked)="_onBackButtonClicked()"
         (tradeExecuted)="_onTradeExecuted()"
@@ -79,6 +81,7 @@
     <div *ngIf="screenToShow == this.TRADE_CREATOR_COMPLETE_SCREEN">
       <trade-creator-complete
         [creatorCoinTrade]="creatorCoinTrade"
+        [userFollowingCreator]="userFollowingCreator"
         (tradeAgainButtonClicked)="_onBackButtonClicked()"
       ></trade-creator-complete>
     </div>
